Double-check that staged file still exists before moving in place
We are running multiple polling tasks to move staged files in place, but
we seem to not ensure that only one task processes a given stage
request which produces lots of benign file-does-not-exist errors.

Double-check that the staged file still exists after sleeping in wait
for dsmc to set file attributes, it's not atomic but should be much
better than not checking at all.