Rework staging and the watching provider to handle real load
First, don't lock thread while waiting for staging inFile to complete.
Doing Thread.sleep locks the thread, it's better to allow it to
reschedule and do other work. This explains why the single-threaded
watching provider fell apart with real load, and also the
1-Hz-per-thread performance of the polling provider.  As a part of this
the default polling period was modified to sync up with the 1000 ms
GRACE_PERIOD.

Second, split staging in two phases, first to detect that a staged file
exists and second to wait for staging completion.

This allows us to do allocation as soon as the file is being staged
instead of waiting until staging is complete before allocating space for
it.

This also enabled sorting out the Watching provider to actually work
with real load. The main problem here was the use of MODIFY events which
actually reports an event for each write operation to a file in the
watched directory, this essentially produces an event storm at the IO
rates we need in production. To add insult to injury, the Java event
implementation doesn't support the inotify IN_CLOSE_WRITE event so we
need to resort to polling instead to monitor file completion.

While at it, use a timeout when watching events to double-check that
we're in sync with the state of the watched directory since this
mechanism isn't foolproof in some corner cases.

The watching provider has now gained the same threads and period
tunables as the polling provider, but with a smaller period default
since the watching provider uses polling only on files that are in
flight.

There is also a new watchtimeout that defaults to 300 seconds that is
used to trigger the double-check mentioned above.