Make graceperiod configureable
Make the grace period between detecting file complete and moving to
destination configureable.

The purpose of the grace period is to allow for the stage operation to
complete before moving the file to the final destination, typically dsmc
needs to set file permissions and timestamps before it considers the
operation completed.

This allows environments not using the original ENDIT daemons/dsmc
combination to tune or disable as needed for that particular setup.